requirements.txt
accelerate==0.26.1
langchain==0.1.1
langchain-openai==0.0.3
matplotlib==3.7.4
notebook==7.0.7
openai==1.9.0
pandas==2.0.3
python-dotenv==1.0.0
scikit-learn==1.3.2
seaborn==0.13.1
SQLAlchemy==2.0.25
torch==2.1.2
transformers==4.36.2
